--- a/setup.py
+++ b/setup.py
@@ -31,6 +31,7 @@ dependencies = [
     "watchdog==2.1.6",  # Filesystem event watching - watches keyring.yaml
     "dnslib==0.9.14",  # dns lib
     "typing-extensions==4.0.1",  # typing backports like Protocol and TypedDict
+    "zstd==1.5.0.4",
 ]
 
 upnp_dependencies = [