@mariano54 2021-12-06
I like the change and running over the whole block seems fine, since the generator is the largest piece anyway. However I don't like adding new dependencies. Have you looked at the python standard lib?

https://docs.python.org/3/library/archiving.html

Can we use one of those algorithms instead?