@mariano54 2021-12-06
It seems like it's much faster than zlib:
https://facebook.github.io/zstd/

However zlib is still very fast and already included in python standard library. Also it's not clear from the website, zstd is lossless compression right?