@arvidn 2021-12-06
> However zlib is still very fast and already included in python standard library. Also it's not clear from the website, zstd is lossless compression right?

yes