@arvidn 2021-12-09
in order to support two (or more) compression algorithms, the blob should also have a field indicating which one was used. I think being able to copy someone else's database and just plop it onto your node is a good feature. If the target node doesn't support the `zstd`, it wouldn't work, but we could at least make it have a human understandable error message.