@arvidn 2021-12-09
Also, I don't think this should land until I run the extended banchmark, to get a better picture of its impact. The benchmark I ran so far only measured time of inserting a new block, not to do lookups.