@arvidn 2021-12-09
ok, I've updated the description with new benchmarks. It still looks like there are no material performance issues.
How does packaging work? Is it sufficient to declare the dependency in `setup.py` to have the module built and included when building packages?