@emlowe 2021-12-10
My understanding re: packaging is that yes, listing in setup.py is sufficient as pyinstaller and the other tools just then work from that.